fix inaccurate types
